# api/openapi.yaml
openapi: 3.1.0
info:
  title: ToolHub Browser API
  version: 2.0.0
  description: |
    Generation-2 single-user control-plane API. Unsafe authenticated requests
    require the session CSRF token. Operations that declare Idempotency-Key
    reject key reuse with a different request.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {$ref: '#/components/responses/Session'}
        '401': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/Error'}
...
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '403': {$ref: '#/components/responses/Error'}
...
  /account/username:
    patch:
      tags: [Auth]
      operationId: updateUsername
      description: Changes the singleton username and revokes every session.
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [username, currentPassword]
              properties:
                username: {$ref: '#/components/schemas/Username'}
                currentPassword: {type: string, format: password, writeOnly: true}
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '400': {$ref: '#/components/responses/Error'}
  /account/password:
    patch:
      tags: [Auth]
      operationId: updatePassword
      description: Changes the singleton password and revokes every session.
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [currentPassword, newPassword]
              properties:
                currentPassword: {type: string, format: password, writeOnly: true}
                newPassword: {type: string, format: password, minLength: 12, maxLength: 1024, writeOnly: true}
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '400': {$ref: '#/components/responses/Error'}
...
  /skills/upload:
    post:
      tags: [Library]
      operationId: uploadSkill
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties: {file: {type: string, format: binary}}
      responses:
        '200': {$ref: '#/components/responses/Skill'}
        '201': {$ref: '#/components/responses/Skill'}
        '400': {$ref: '#/components/responses/Error'}
  /skills/import:
    post:
      tags: [Library]
      operationId: importSkill
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SkillSourceInput'}}}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
  /market/search:
    get:
      tags: [Library]
      operationId: searchMarketplace
      parameters:
        - {name: q, in: query, required: true, schema: {type: string, minLength: 2, maxLength: 200}}
        - {name: source, in: query, schema: {type: string, enum: [skillsmp, xiaping]}}
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200': {description: Normalized provider result, content: {application/json: {schema: {type: object}}}}
        '429': {$ref: '#/components/responses/Error'}
        '502': {$ref: '#/components/responses/Error'}
  /recommendations:
    post:
      tags: [Library]
      operationId: recommendSkills
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {$ref: '#/components/requestBodies/JSONObject'}
      responses:
        '200':
          description: Recommendations; automaticInstall is always false.
          content:
            application/json:
              schema:
                type: object
                required: [candidates, automaticInstall]
                properties:
                  candidates: {type: array, items: {type: object}}
                  automaticInstall: {const: false}
  /mcp/servers:
    get:
      tags: [Library]
      operationId: listMCPServers
      responses:
        '200': {$ref: '#/components/responses/MCPServerList'}
    post:
      tags: [Library]
      operationId: createMCPServer
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/MCPServerInput'}}}
      responses:
        '201': {$ref: '#/components/responses/MCPServer'}
        '400': {$ref: '#/components/responses/Error'}
  /mcp/servers/{serverID}:
    parameters: [{$ref: '#/components/parameters/ServerID'}]
    put:
      tags: [Library]
      operationId: updateMCPServer
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/MCPServerInput'}}}
      responses:
        '200': {$ref: '#/components/responses/MCPServer'}
        '400': {$ref: '#/components/responses/Error'}
    delete:
      tags: [Library]
      operationId: deleteMCPServer
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '409': {$ref: '#/components/responses/Error'}
  /mcp/import:
    post:
      tags: [MCP]
      operationId: importLocalMCPServer
      description: Consumes one revision-bound confirmation and queues an ephemeral, one-time local MCP capture.
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmationToken]
              properties: {confirmationToken: {type: string, minLength: 1, writeOnly: true}}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
        '409': {$ref: '#/components/responses/Error'}
  /profiles:
    get:
      tags: [Profiles]
      operationId: listProfiles
      responses:
        '200': {$ref: '#/components/responses/ProfileList'}
    post:
      tags: [Profiles]
      operationId: createProfile
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ProfileInput'}}}
      responses:
        '201': {$ref: '#/components/responses/Profile'}
  /profiles/{profileID}:
    parameters: [{$ref: '#/components/parameters/ProfileID'}]
    put:
      tags: [Profiles]
      operationId: updateProfile
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ProfileInput'}}}
      responses:
        '200': {$ref: '#/components/responses/Profile'}
        '409': {$ref: '#/components/responses/Error'}
    delete:
      tags: [Profiles]
      operationId: deleteProfile
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '409': {$ref: '#/components/responses/Error'}
  /profiles/{profileID}/preflight:
    post:
      tags: [Profiles]
      operationId: preflightProfile
      description: Resolves current Library revisions and returns five-minute, one-use confirmation tokens.
      parameters:
        - {$ref: '#/components/parameters/ProfileID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [targetIds]
              properties:
                targetIds: {type: array, minItems: 1, maxItems: 100, uniqueItems: true, items: {$ref: '#/components/schemas/UUID'}}
      responses:
        '200': {$ref: '#/components/responses/PreflightList'}
        '409': {$ref: '#/components/responses/Error'}
  /profiles/{profileID}/apply:
    post:
      tags: [Profiles]
      operationId: applyProfile
      parameters:
        - {$ref: '#/components/parameters/ProfileID'}
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmationTokens]
              properties:
                confirmationTokens: {type: array, minItems: 1, maxItems: 100, uniqueItems: true, items: {type: string}}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
        '409': {$ref: '#/components/responses/Error'}
...
  /nodes/refresh:
    post:
      tags: [Targets]
      operationId: refreshNodes
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
  /nodes/{nodeID}:
    patch:
      tags: [Targets]
      operationId: updateNodeManagedUsername
      description: Sets a Salt node managed-username override. An empty string clears the override and restores the global managed username.
      parameters:
        - {$ref: '#/components/parameters/NodeID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [managedUsername]
              properties: {managedUsername: {type: string, maxLength: 32, pattern: '^(?:[a-z_][a-z0-9_-]{0,31})?$'}}
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
...
  /targets/{targetID}:
    get:
      tags: [Targets]
      operationId: getTarget
      parameters: [{$ref: '#/components/parameters/TargetID'}]
      responses:
        '200':
          description: Target, latest inventory, and active immutable desired snapshot.
          content: {application/json: {schema: {$ref: '#/components/schemas/TargetDetail'}}}
        '404': {$ref: '#/components/responses/Error'}
  /targets/{targetID}/scan:
    post:
      tags: [Targets]
      operationId: scanTarget
      parameters:
        - {$ref: '#/components/parameters/TargetID'}
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
  /targets/{targetID}/skill-import:
    post:
      tags: [Targets]
      operationId: importLocalSkill
      description: Queues an immutable Library import of one scanned, unmanaged local Skill.
      parameters:
        - {$ref: '#/components/parameters/TargetID'}
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, expectedRevision, contentHash]
              properties:
                name: {type: string, minLength: 1}
                expectedRevision: {$ref: '#/components/schemas/SHA256'}
                contentHash: {$ref: '#/components/schemas/SHA256'}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
        '409': {$ref: '#/components/responses/Error'}
  /targets/{targetID}/mcp-import/preflight:
    post:
      tags: [Targets]
      operationId: preflightLocalMCPImport
      description: Returns only sanitized local user-scope MCP fields and key names; secret values are never returned to the browser.
      parameters:
        - {$ref: '#/components/parameters/TargetID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false}}}
      responses:
        '200':
          description: Five-minute, one-use confirmations bound to the current native configuration revision.
          content: {application/json: {schema: {$ref: '#/components/schemas/LocalMCPImportPreflight'}}}
        '409': {$ref: '#/components/responses/Error'}
  /targets/{targetID}/edit:
    post:
      tags: [Targets]
      operationId: editTarget
      parameters:
        - {$ref: '#/components/parameters/TargetID'}
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedRevision, skillIds, mcpServerIds]
              properties:
                expectedRevision: {$ref: '#/components/schemas/SHA256'}
                skillIds: {type: array, maxItems: 500, uniqueItems: true, items: {$ref: '#/components/schemas/UUID'}}
                mcpServerIds: {type: array, maxItems: 500, uniqueItems: true, items: {$ref: '#/components/schemas/UUID'}}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
  /targets/{targetID}/backups:
    get:
      tags: [Targets]
      operationId: listBackups
      parameters: [{$ref: '#/components/parameters/TargetID'}]
      responses:
        '200': {$ref: '#/components/responses/BackupList'}
  /targets/{targetID}/restore:
    post:
      tags: [Targets]
      operationId: restoreTarget
      parameters:
        - {$ref: '#/components/parameters/TargetID'}
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [backupId, expectedRevision]
              properties: {backupId: {$ref: '#/components/schemas/UUID'}, expectedRevision: {$ref: '#/components/schemas/SHA256'}}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
  /targets/{targetID}/relay/{action}:
    post:
      tags: [Targets]
      operationId: relayAction
      parameters:
        - {$ref: '#/components/parameters/TargetID'}
        - {name: action, in: path, required: true, schema: {type: string, enum: [start, stop, restart, health]}}
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
...
  /operations/{operationID}:
    get:
      tags: [Operations]
      operationId: getOperation
      parameters: [{$ref: '#/components/parameters/OperationID'}]
      responses:
        '200': {$ref: '#/components/responses/Operation'}
        '404': {$ref: '#/components/responses/Error'}
  /operations/{operationID}/cancel:
    post:
      tags: [Operations]
      operationId: cancelOperation
      parameters:
        - {$ref: '#/components/parameters/OperationID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '409': {$ref: '#/components/responses/Error'}
  /operations/{operationID}/retry-failed:
    post:
      tags: [Operations]
      operationId: retryFailedTargets
      parameters:
        - {$ref: '#/components/parameters/OperationID'}
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
...
  /settings:
    get:
      tags: [Settings]
      operationId: getSettings
      responses:
        '200': {$ref: '#/components/responses/Settings'}
    put:
      tags: [Settings]
      operationId: updateSettings
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SettingsInput'}}}
      responses:
        '200': {$ref: '#/components/responses/Settings'}
        '400': {$ref: '#/components/responses/Error'}
  /updates/check:
    post:
      tags: [Settings]
      operationId: checkUpdates
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '202': {$ref: '#/components/responses/Operation'}
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: toolhub_session}
  parameters:
    CSRFToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 200}}
    NodeID: {name: nodeID, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
    TargetID: {name: targetID, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
    ProfileID: {name: profileID, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
    ServerID: {name: serverID, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
    OperationID: {name: operationID, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
  requestBodies:
    JSONObject:
      required: true
      content: {application/json: {schema: {type: object}}}
  responses:
    NoContent: {description: No content}
    Error:
      description: Error envelope
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Session: {description: Session state, content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}}
...
    Skill: {description: Immutable Skill library record, content: {application/json: {schema: {$ref: '#/components/schemas/Skill'}}}}
...
    MCPServer: {description: MCP server without secret values, content: {application/json: {schema: {$ref: '#/components/schemas/MCPServer'}}}}
    MCPServerList: {description: MCP servers, content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/MCPServer'}}}}}}}
    Profile: {description: Unified Profile, content: {application/json: {schema: {$ref: '#/components/schemas/Profile'}}}}
    ProfileList: {description: Profiles, content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Profile'}}}}}}}
...
    Operation: {description: Durable operation, content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}}
...
    BackupList: {description: Backups, content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Backup'}}}}}}}
...
    Settings: {description: Singleton settings, content: {application/json: {schema: {$ref: '#/components/schemas/Settings'}}}}
    PreflightList: {description: Per-target preflight confirmations, content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/PreflightItem'}}}}}}}
  schemas:
    UUID: {type: string, format: uuid}
    Username: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{2,31}$'}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
            details: {type: object}
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {$ref: '#/components/schemas/Username'}
        password: {type: string, format: password, writeOnly: true}
...
    Session:
      type: object
      additionalProperties: false
      required: [authenticated, user, csrfToken, expiresAt]
      properties:
        authenticated: {const: true}
        user: {$ref: '#/components/schemas/Account'}
        csrfToken: {type: string, writeOnly: true}
        expiresAt: {type: string, format: date-time}
...
    SkillSourceInput:
      type: object
      additionalProperties: false
      required: [kind, name]
      properties:
        kind: {type: string, enum: [git, skillsmp, xiaping]}
        name: {type: string}
        url: {type: string, format: uri}
        subdirectory: {type: string}
        commit: {type: string}
        metadata: {type: object}
    Skill:
      type: object
      required: [id, slug, name, sourceKind, currentVersionId, currentSha256, currentContentHash, manifest, scanReport, createdAt, updatedAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        slug: {type: string}
        name: {type: string}
        description: {type: string}
        sourceKind: {type: string, enum: [zip, git, skillsmp, xiaping, local]}
        sourceUrl: {type: string}
        sourceCommit: {type: string}
        currentVersionId: {$ref: '#/components/schemas/UUID'}
        currentSha256: {$ref: '#/components/schemas/SHA256'}
        currentContentHash: {$ref: '#/components/schemas/SHA256'}
        manifest: {type: object}
        scanReport: {type: object}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    MCPServerInput:
      type: object
      additionalProperties: false
      required: [name, description, transport, command, args, url, env, headers]
      properties:
        name: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{0,127}$'}
        description: {type: string}
        transport: {type: string, enum: [stdio, http, sse]}
        command: {type: string}
        args: {type: array, items: {type: string}}
        url: {type: string}
        env: {$ref: '#/components/schemas/WriteOnlySecretMap'}
        headers: {$ref: '#/components/schemas/WriteOnlySecretMap'}
    WriteOnlySecretMap:
      type: object
      writeOnly: true
      additionalProperties: {type: string, writeOnly: true}
      description: Non-empty values create/replace secrets; an existing key with an empty value is retained; an omitted key is removed.
    MCPServer:
      type: object
      required: [id, name, revision, transport, args, envKeys, headerKeys, contentHash, createdAt, updatedAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        name: {type: string}
        description: {type: string}
        revision: {type: integer, minimum: 1}
        transport: {type: string, enum: [stdio, http, sse]}
        command: {type: string}
        args: {type: array, items: {type: string}}
        url: {type: string}
        envKeys: {type: array, items: {type: string}}
        headerKeys: {type: array, items: {type: string}}
        contentHash: {$ref: '#/components/schemas/SHA256'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    LocalMCPServerPreview:
      type: object
      additionalProperties: false
      required: [name, transport, args, envKeys, headerKeys, contentHash, confirmationToken, expiresAt]
      properties:
        name: {type: string}
        transport: {type: string, enum: [stdio, http, sse]}
        command: {type: string}
        args: {type: array, items: {type: string}}
        url: {type: string}
        envKeys: {type: array, items: {type: string}}
        headerKeys: {type: array, items: {type: string}}
        contentHash: {$ref: '#/components/schemas/SHA256'}
        confirmationToken: {type: string, writeOnly: true}
        expiresAt: {type: string, format: date-time}
    LocalMCPImportPreflight:
      type: object
      additionalProperties: false
      required: [targetRevision, items]
      properties:
        targetRevision: {$ref: '#/components/schemas/SHA256'}
        items: {type: array, items: {$ref: '#/components/schemas/LocalMCPServerPreview'}}
    ProfileInput:
      type: object
      additionalProperties: false
      required: [name, description, revision, skillIds, mcpServerIds]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        description: {type: string}
        revision: {type: integer, minimum: 0}
        skillIds: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/UUID'}}
        mcpServerIds: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/UUID'}}
    Profile:
      allOf:
        - {$ref: '#/components/schemas/ProfileInput'}
        - type: object
          required: [id, createdAt, updatedAt]
          properties:
            id: {$ref: '#/components/schemas/UUID'}
            createdAt: {type: string, format: date-time}
            updatedAt: {type: string, format: date-time}
...
    Target:
      type: object
      required: [id, targetKey, nodeId, nodeName, nodeKind, runtime, managedUsername, writable, health, desiredRevision]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        targetKey: {type: string}
        nodeId: {$ref: '#/components/schemas/UUID'}
        nodeName: {type: string}
        nodeKind: {type: string, enum: [local, salt]}
        saltMinionId: {type: string}
        runtime: {type: string, enum: [claude, codex, hermes, shared-relay]}
        managedUsername: {type: string}
        writable: {type: boolean}
        health: {$ref: '#/components/schemas/Health'}
        desiredRevision: {type: integer, minimum: 0}
        targetRevision: {type: string}
        driftSummary: {type: object}
        lastScannedAt: {type: [string, 'null'], format: date-time}
        lastReconciledAt: {type: [string, 'null'], format: date-time}
        errorCode: {type: string}
        errorReason: {type: string}
    TargetDetail:
      type: object
      required: [target, inventory, targetRevision]
      properties:
        target: {$ref: '#/components/schemas/Target'}
        inventory:
          type: object
          properties:
            members: {type: array, items: {$ref: '#/components/schemas/InventoryMember'}}
            relay: {$ref: '#/components/schemas/RelayStatus'}
        targetRevision: {type: string}
        desired:
          oneOf:
            - type: 'null'
            - type: object
              required: [snapshot, manifest]
              properties:
                snapshot: {$ref: '#/components/schemas/DesiredSnapshot'}
                manifest: {$ref: '#/components/schemas/DesiredManifest'}
    InventoryMember:
      type: object
      required: [id, kind, name, protected]
      properties:
        id: {type: string}
        kind: {type: string, enum: [skill, mcp, anchor, entry]}
        name: {type: string}
        contentHash: {type: string}
        protected: {type: boolean}
        scope: {type: string}
        revision: {type: integer, minimum: 1}
        secretKeys: {type: array, items: {type: string}}
    DesiredManifest:
      type: object
      additionalProperties: false
      required: [schemaVersion, target, skills, mcpServers, managedMemberIds]
      properties:
        schemaVersion: {const: 1}
        target: {$ref: '#/components/schemas/ManifestTarget'}
        profileId: {$ref: '#/components/schemas/UUID'}
        profileRevision: {type: integer, minimum: 1}
        skills: {type: array, items: {$ref: '#/components/schemas/SkillMember'}}
        mcpServers: {type: array, items: {$ref: '#/components/schemas/MCPMember'}}
        managedMemberIds: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/UUID'}}
        relayPort: {type: integer, minimum: 1, maximum: 65535}
...
    DesiredSnapshot:
      type: object
      required: [id, revision, sourceKind, manifestHash, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        revision: {type: integer, minimum: 1}
        sourceKind: {type: string, enum: [profile_apply, target_edit, restore]}
        sourceId: {$ref: '#/components/schemas/UUID'}
        profileRevision: {type: integer}
        manifestHash: {$ref: '#/components/schemas/SHA256'}
        createdAt: {type: string, format: date-time}
...
    PreflightItem:
      type: object
      required: [targetId, confirmationToken, expiresAt, result]
      properties:
        targetId: {$ref: '#/components/schemas/UUID'}
        confirmationToken: {type: string, writeOnly: true}
        expiresAt: {type: string, format: date-time}
        result:
          type: object
          required: [targetRevision, manifestHash, diff]
          properties:
            targetRevision: {type: string}
            manifestHash: {$ref: '#/components/schemas/SHA256'}
            diff: {$ref: '#/components/schemas/Diff'}
    Operation:
      type: object
      required: [id, kind, status, metadata, cancelRequested, createdAt, updatedAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        kind: {type: string, enum: [skill_import, mcp_import, update_check, apply, edit, restore, scan, refresh, reconcile, relay_start, relay_stop, relay_restart, relay_health, backup_gc]}
        status: {$ref: '#/components/schemas/OperationStatus'}
        sourceId: {$ref: '#/components/schemas/UUID'}
        metadata: {type: object}
        errorCode: {type: string}
        errorReason: {type: string}
        cancelRequested: {type: boolean}
        targets: {type: array, items: {$ref: '#/components/schemas/OperationTarget'}}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
        updatedAt: {type: string, format: date-time}
...
    SettingsInput:
      type: object
      additionalProperties: false
      required: [managedUsername, updateCron, timezone, relayPort, relayIntentionalPaused]
      properties:
        managedUsername: {type: string, pattern: '^[a-z_][a-z0-9_-]{0,31}$'}
        updateCron: {type: string}
        timezone: {type: string}
        relayPort: {type: integer, minimum: 1, maximum: 65535}
        relayIntentionalPaused: {type: boolean}
        updatedAt: {type: string, format: date-time, readOnly: true}
...
    RelayStatus:
      type: object
      required: [state, healthy, intentionalPaused, endpoint]
      properties:
        state: {type: string}
        healthy: {type: boolean}
        intentionalPaused: {type: boolean}
        endpoint: {type: string, format: uri}
        version: {type: string}
        errorCode: {type: string}
    SHA256: {type: string, pattern: '^[a-f0-9]{64}$'}
...